% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{Buildct}
\alias{Buildct}
\title{Tree size distrib to BA of a plot}
\usage{
Buildct(mesh, SurfEch = 0.03)
}
\arguments{
\item{mesh}{Vector of all values between L and U of the IPM
(L being the smallest size and U the largest.). Length of this vector is
the number of size class in the IPM. num.}

\item{SurfEch}{Value of plot size surface in m^2}
}
\value{
numeric vector of basal area computed depending on plot size
}
\description{
Build the vector to pass from a tree size distribution to
a basal area computed for a given plot size
}
\examples{
Buildct(seq(10,50, 10), 5e-4)

}
\keyword{internal}
