% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{clim_gradient}
\alias{clim_gradient}
\title{Build climatic gradient}
\usage{
clim_gradient(
  clim_table,
  start_clim = c("opti", "hot", "cold"),
  end_clim = c("opti", "hot", "cold"),
  times = 100,
  noise = rnorm,
  sigma = c(1, 0.001),
  seed = 42,
  extrapolate = 0
)
}
\arguments{
\item{clim_table}{climate table from data climate_species, filtered on a species}

\item{start_clim}{strating climate in opti, hot or cold}

\item{end_clim}{same as start_clim}

\item{times}{duration of the simulation without extrapolation}

\item{noise}{function to draw noise from. Default is rnorm. Shoud have a
n value to draw and an sd parameter.}

\item{sigma}{standard deviation value for the noise, for the two climatic variable
sgdd and wai.}

\item{seed}{for the random effect of the noise function}

\item{extrapolate}{extrapolation of the climatic tendancy in percent of the
times. Example, for 100 year and an extrapolation of 50\% will add 50 years of
simulations beyond the last climate.}
}
\description{
Build climatic gradient
}
\details{
The climate will be linear between the two climate, using a seq() function.
clim_table must encode the climate with numeric 1, 2 and 3 respectively
values for hot, opti and cold.
}
