% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{correction}
\alias{correction}
\alias{correction.ipm}
\alias{correction.species}
\alias{correction.forest}
\title{Add correction states to a system}
\usage{
correction(x, correction = "none")

\method{correction}{ipm}(x, correction = "none")

\method{correction}{species}(x, correction = "none")

\method{correction}{forest}(x, correction = "none")
}
\arguments{
\item{x}{an object that require a correction addition}

\item{correction}{Type of correction, match between cut and none (default).
chr.}
}
\description{
Apply correction to a matrix by setting to 0 all right and bottom values
}
