% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_disturb}
\alias{def_disturb}
\title{Default disturbance function}
\usage{
def_disturb(x, species, disturb = NULL, ...)
}
\arguments{
\item{x}{population state distribution at time t}

\item{species}{The species class object of interest to get mesh and RDIcoef
values from. RDIcoef is a one line dataframe with RDI coefficient for one
species.}

\item{disturb}{Disturbance parameters. Highly depend on the disturbance
impact parameters given to the species.}

\item{...}{Default disturbance function does not require}
}
\description{
Default disturbance function
}
