% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_harv}
\alias{def_harv}
\title{Default population harvest}
\usage{
def_harv(x, species, ...)
}
\arguments{
\item{x}{population state at time t}

\item{species}{ignored}

\item{...}{Variables used in this case of Uneven harvest
\describe{
\item{ct}{is the vector to compute BA with x (ct = Buildct(mesh, SurfEch))}
}}
}
\value{
Distribution of population to harvest.
Values are between 0 (null harvest) and Xi.
}
\description{
Constant rate harvest of 0.06 percent per year
(check if harvest frequence is 1 in forest object).
}
\details{
This function is the default function for \code{harvest_fun} function of
a species and always takes arguments \code{x}, \code{species} plus specific
argument from different harvest models..
}
\seealso{
Other functions that defines harvest rules for a species.: 
\code{\link{Even_harv}()},
\code{\link{Uneven_harv}()},
\code{\link{species}()}
}
\concept{functions that defines harvest rules for a species.}
