% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init_even}
\alias{def_init_even}
\title{Default even population initialization}
\usage{
def_init_even(mesh, SurfEch = 0.03)
}
\arguments{
\item{mesh}{all possible states of a population, based on an IPM.
Minimal and maximal values are respectively U and L, for a total number of
m states.}

\item{SurfEch}{Value of plot size surface in \eqn{m^2}}
}
\description{
The population will initiate with only individual in the 5 first cells of the
mesh.
}
\seealso{
Other functions for initiating species population during simulation: 
\code{\link{def_init}()},
\code{\link{def_initBA}()},
\code{\link{def_init_k}()},
\code{\link{species}()}
}
\concept{functions for initiating species population during simulation}
