% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init_kpop}
\alias{def_init_kpop}
\title{Init population with precise distribution}
\usage{
def_init_kpop(x)
}
\arguments{
\item{x}{Population to draw systematically. This vector should
be composed of values in \code{[-lag, maxdbh]} values.
This is the distribution for the sampled plot.}
}
\value{
Function similar def_init but with no random effect anymore.
}
\description{
Init population with precise distribution
}
\details{
Possible values are :
\describe{
\item{size in \eqn{[1; \inf [}}{diameter in $mm$ for a single tree}
\item{size in \eqn{]-lag; -1]}}{coded value for individual in the lag.
When a new individual is recruited, it will start at the maximum lag
value (-lag) and grow with \eqn{s_{t+1} = s_t + 1}.
When \eqn{s_t = 1}, then \eqn{s_{t+1} = 90}.}
\item{size = 0}{coded value for a dead individual.
These individuals are removed from memory during a simulation.}
}
}
\seealso{
Other functions for initiating species population during simulation: 
\code{\link{def_init}()},
\code{\link{def_initBA}()},
\code{\link{def_initBA_pop}()},
\code{\link{def_init_even}()},
\code{\link{def_init_k}()},
\code{\link{def_init_pop}()},
\code{\link{species}()}
}
\concept{functions for initiating species population during simulation}
