% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init_planting}
\alias{def_init_planting}
\title{Init population from plantation, after even clear cut}
\usage{
def_init_planting(species)
}
\arguments{
\item{species}{name of the species. This allows to extract the parameters
from matreex::distrib_planting dataset, that have been computed by Georges in
issue #13.}
}
\description{
Init population from plantation, after even clear cut
}
\details{
This function may replace def_init_even() in the long term.
}
