% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init_pop}
\alias{def_init_pop}
\title{Default individual population initialization}
\usage{
def_init_pop(mesh, SurfEch = 0.03)
}
\arguments{
\item{mesh}{all possible states of a population, based on an IPM.
Minimal and maximal values are respectively U and L, for a total number of
m states.}

\item{SurfEch}{Value of plot size surface in \eqn{m^2}}
}
\description{
The population will initiate with a random distribution to match a basal area
of 1.
}
\seealso{
Other functions for initiating species population during simulation: 
\code{\link{def_init}()},
\code{\link{def_initBA}()},
\code{\link{def_initBA_pop}()},
\code{\link{def_init_even}()},
\code{\link{def_init_k}()},
\code{\link{def_init_kpop}()},
\code{\link{species}()}
}
\concept{functions for initiating species population during simulation}
