% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{delay.dsCMatrix}
\alias{delay.dsCMatrix}
\title{Delay dsCMatrix}
\usage{
\method{delay}{dsCMatrix}(x, delay = 0)
}
\arguments{
\item{x}{an object that require a delay addition}

\item{delay}{the number of time delay to add. single int. Its possible to
give negative values to remove delay from previously delayed objects. However
final delay must be positive (0 is acceptable).}
}
\description{
Adding a topleft corner to a matrix filled with 0.
}
\examples{
x <- new("dtCMatrix", i = c(0L, 1L, 2L, 1L, 2L, 2L), p = c(0L, 3L,  5L, 6L),
         Dim = c(3L, 3L), x = c(0.1, 0.2, 0.3, 0.1, 0.2, 0.1), uplo = "L", diag = "N")
x <- delay(x, 2)
delay(x, -1)
}
