% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{delay.dtCMatrix}
\alias{delay.dtCMatrix}
\title{Delay dtCMatrix}
\usage{
\method{delay}{dtCMatrix}(x, delay = 0)
}
\arguments{
\item{x}{an object that require a delay addition}

\item{delay}{the number of time delay to add. single positive int.}
}
\description{
Adding a topleft corner to a matrix filled with 0.
}
\examples{
x <- new("dtCMatrix", i = c(0L, 1L, 2L, 1L, 2L, 2L), p = c(0L, 3L,  5L, 6L),
         Dim = c(3L, 3L), x = c(1, 2, 3, 1, 2, 1), uplo = "L", diag = "N")
delay(x, 2)
}
