% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_planting.R
\docType{data}
\name{distrib_planting}
\alias{distrib_planting}
\title{Species distribution parameter for young stands.}
\format{
An object of class \code{data.frame} with 45 rows and 6 columns.
}
\source{
Georges Kunstler
}
\usage{
distrib_planting
}
\description{
Species distribution parameter for young stands.
}
\keyword{datasets}
