% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disturb_coef.R
\docType{data}
\name{disturb_coef}
\alias{disturb_coef}
\title{Species disturbance parameters.}
\format{
An object of class \code{data.frame} with 64 rows and 10 columns.
}
\source{
https://doi.org/10.1111/gcb.16630
}
\usage{
disturb_coef
}
\description{
Species disturbance parameters.
}
\keyword{datasets}
