% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{Even_harv}
\alias{Even_harv}
\title{Perform the harvest Even}
\usage{
Even_harv(x, species, ...)
}
\arguments{
\item{x}{population state distribution at time t}

\item{species}{The species class object of interest to get mesh and RDIcoef
values from. RDIcoef is a one line dataframe with RDI coefficient for one
species.}

\item{...}{Variables used in this case of Uneven harvest
\describe{
\item{targetKg}{is the target Kg}
\item{targetRDI}{is the target RDI}
\item{}
}}
}
\description{
Perform the harvest Even
}
