% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_indiv.R
\name{extinct}
\alias{extinct}
\title{Extract extinction time for a simulation.}
\usage{
extinct(sim, all = TRUE)
}
\arguments{
\item{sim}{Simulation output from sim_indiv_forest}

\item{all}{TRUE will return single value for all species, FALSE return named
vector with the value for each species.}
}
\value{
time at which populations were extinct. If the simulation ended
with any(N > 0), returns NA.
}
\description{
Extinction time is time when there is a total population of 0,
for all species bu default.
}
