% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fit_sgr.R
\name{fit_sgr}
\alias{fit_sgr}
\title{Create a new fitted models from}
\usage{
fit_sgr(
  sv_params,
  sv_family,
  gr_params,
  gr_sigma,
  rec_params,
  species,
  max_dbh
)
}
\arguments{
\item{sv_params}{Named vector of survival parameters fitted for this species
and climatic condition. Minimal parameters are intercept and size.}

\item{sv_family}{Family list contain the details of the models used when
fitting by functions such as glm. Only the inverse of the link function
(linkinv) is used in the model for now. list.}

\item{gr_params}{Named vector of growth parameters fitted for this species and
climatic condition. Minimal parameters are intercept and size.}

\item{gr_sigma}{Standard deviation of the residuals for the growth fitted
model.}

\item{rec_params}{Named vector of growth parameters fitted for this species and
climatic condition. Minimal parameters are intercept, BATOTSP and BATOTNonSP.}

\item{species}{Name of the species to run simulation on. Single char.}

\item{max_dbh}{Maximum diameter of the fitted dataset. Single dbh.}
}
\description{
Species are defined by an IPM which is a transition matrix from size between
t and t+1, recruitment and harvest functions. Each species has these items
defined for a given climate.
An additionnal vector of harvest parameers is required with minimal size to
harvest (dth), size above wich harvest is constant (dha).
}
