% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_forest.R
\name{forest}
\alias{forest}
\title{Create a new forest for simulation}
\usage{
forest(
  species = list(),
  harv_rules = c(Pmax = 0.25, dBAmin = 3, freq = 1, alpha = 1)
)
}
\arguments{
\item{species}{List of species created with treeforce package.}

\item{harv_rules}{Vector for harvest rules at the scale of the forest.
\describe{
\item{Pmax}{maximum proportion of BAcut / BA}
\item{dBAmin}{the minimum BA to perform cut}
\item{freq}{Frequence at which the harvest will be executed.}
}}
}
\description{
A forest is a group of one of multiple species to silumate along time using
the IPM defined for each species and harvest rules.
}
