% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{get_maxdbh}
\alias{get_maxdbh}
\title{Get the max_dbh of a fitted species}
\usage{
get_maxdbh(x)
}
\arguments{
\item{x}{treeforce class object. Used on fit_sgr for now.}
}
\description{
Get the max_dbh of a fitted species
}
