% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_IPM.R
\name{get_step_IPM}
\alias{get_step_IPM}
\title{Extract an IPM matrix}
\usage{
get_step_IPM(x, ...)
}
\arguments{
\item{x}{IPM or mu_gr class object.}

\item{...}{Variables used depending on the class of x.
\describe{
\item{BA}{Total basal area to get the IPM for during simulation.}
\item{climate}{Climate for which the IPM is needed. Only used for mu_gr.}
\item{sim_corr}{Simulation correction applied to the IPM. "cut" or "none"}
}}
}
\description{
Extract the IPM matrix needed during simulation for a given BA and optionally
climate.
}
\details{
Methods are set for IPM objects and mu_gr. Only the second one require
to give the climate and simulation correction to apply to the matrix.
}
