% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mu_gr.R
\name{getRangemu}
\alias{getRangemu}
\title{Get mu range for a species}
\usage{
getRangemu(climate, fit, BA = 0:200, mesh = seq(90, 900, by = 10))
}
\arguments{
\item{climate}{climate table for the species.}

\item{fit}{Fitted model for growth and survival of the species and climate
given. Functions will depend on size and basal area.}

\item{BA}{Vector of basal area to integrate on. Integrating on 0 is important
so use it. Integrating above 200 is absurd.}

\item{mesh}{vector of mesh variables. m is the number of bins, L is the
minimum size and U the maximum size. h will be defined in the function as
\eqn{h <- (U - L) / m}.}
}
\description{
Get mu range for a species
}
\keyword{internal}
