% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_in.R
\name{make_mutrix}
\alias{make_mutrix}
\title{Build IPM for a given species and climate}
\usage{
make_mutrix(
  species,
  fit,
  mesh = c(m = 700, L = 90, U = 1500),
  BA = 0:200,
  correction = c("constant", "none", "ceiling", "sizeExtremes"),
  stepMu = 0.001,
  level = c(3, 140),
  diag_tresh = 50,
  midbin_tresh = 25,
  mid_level = 5,
  year_delta = 1,
  IsSurv = TRUE,
  verbose = FALSE
)
}
\description{
Integrate IPM for growth and survival function at a specific climate for a
species on a basal area variation.
}
\details{
The check between climate variables and fitted variable will assert if all
variables in the model are provided expect variables derived from "size"
(size, size2, logsize), "intercept" and "BATOTcomp". An error will be
triggered if the climate variable is missing.

One can desactivate each kind of integration with some treshold values.
A negative value in diag_tresh (ex: -1) will cancel the Gauss-Legendre
integration and a midbin_tresh null value (ex: 0) will cancel the midbin
integration.

This is a working function to test faster integration but it integrate a
ba value
}
\keyword{internal}
