% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_forest.R
\name{new_forest}
\alias{new_forest}
\title{Constructor for forest class}
\usage{
new_forest(
  species = list(),
  harv_rules = c(Pmax = 0.25, dBAmin = 3, freq = 1, alpha = 1),
  favoured_sp = c()
)
}
\arguments{
\item{species}{List of species created with matreex package.}

\item{harv_rules}{Vector for harvest rules at the scale of the forest.
\describe{
\item{Pmax}{maximum proportion of BAcut / BA}
\item{dBAmin}{the minimum BA to perform cut}
\item{freq}{Frequence at which the harvest will be executed.}
\item{alpha}{parameter for uneven harvest, allowing to target more cuts
for abundant species or not. See Harvesting vignette.}
}}

\item{favoured_sp}{Logical named vector to tell if species are favoured during
Uneven harvesting or not. If not NULL, the species names should be the same as
in the species list.}
}
\description{
Only used in the matreex package
}
\keyword{internal}
