% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ipm.R
\name{new_ipm}
\alias{new_ipm}
\title{Constructor for IPM class}
\usage{
new_ipm(
  IPM,
  BA,
  mesh,
  species,
  correction,
  survival = TRUE,
  climatic,
  clim_lab,
  fit,
  compress = FALSE,
  delay = 0,
  int
)
}
\arguments{
\item{IPM}{list of the IPM for different basal area for the species of
interest. List of dtCMatrix.}

\item{BA}{values of BA defined for the IPM, with the same names as IPM. num.}

\item{mesh}{Vector of all values between L and U of the IPM
(L being the smallest size and U the largest.). Length of this vector is
the number of size class in the IPM. num.}

\item{species}{Name of the species to run simulation on. Single char.}

\item{correction}{IPM correction wanted for this species. Single char.}

\item{survival}{IPM inclusion of survival function. Lgl, TRUE by default.}

\item{climatic}{Vector of named climatic values used to fit the ipm on. dbl.}

\item{clim_lab}{Label for climatic used. This values will be matched when
simulating multiple species together.}

\item{fit}{fit_sgr class object that contains parameters for survival,
growth and recruitment.}

\item{compress}{Is the IPM matrix compressed as integer (via \code{x * 1e7}).
Help to limit  size when saved on disc. FALSE by default. lgl.}

\item{delay}{Number of year delay between the recruitment of an individual
and it's inclusion in the IPM. This will enlarge the IPM and add sub diagonal
values of 1. See \code{\link[matreex]{delay}}.}

\item{int}{Internal vector used to store logs about integratio process
that register year delta, maximum integration error on the first half of
the mesh, minimal gauss-legendre value and maximal midbin error.}
}
\description{
Constructor for IPM class
}
