% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mu_gr.R
\name{new_mu_gr}
\alias{new_mu_gr}
\title{Constructor of mu_gr class}
\usage{
new_mu_gr(
  mu_gr,
  BA,
  mesh,
  mu_tab,
  mu_step,
  fit,
  species,
  correction,
  surv = TRUE,
  int
)
}
\arguments{
\item{mu_gr}{Matrix of computed mesh values along a mu gradient.}

\item{BA}{values of BA used to get the range of mu for this species. num.}

\item{mesh}{vector of mesh variables. m is the number of bins, L is the
minimum size and U the maximum size. h will be defined in the function as
\eqn{h <- (U - L) / m}.}

\item{mu_tab}{Mu values for which the values are computed. Each row of mu_gr
matrix is for a given mu.}

\item{fit}{fit_sgr model for survival growth and recruitment for the given
species.}

\item{species}{Name of the species to run simulation on. Single char.}

\item{correction}{IPM correction wanted for this species. Single char.}

\item{int}{Integration information about levels and dimension of integration.}
}
\description{
Constructor of mu_gr class
}
\keyword{internal}
