% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_in.R
\name{new_mutrix}
\alias{new_mutrix}
\title{Constructor of mutrix class}
\usage{
new_mutrix(
  mutrix,
  BA,
  mesh,
  mu_tab,
  mu_step,
  fit,
  species,
  correction,
  surv = TRUE,
  int
)
}
\arguments{
\item{mutrix}{Matrix of computed mesh values along a mu gradient.}

\item{BA}{values of BA used to get the range of mu for this species. num.}

\item{mesh}{vector of mesh variables. m is the number of bins, L is the
minimum size and U the maximum size. h will be defined in the function as
\eqn{h <- (U - L) / m}.}

\item{mu_tab}{Mu values for which the values are computed. Each row of mutrix
matrix is for a given mu.}

\item{fit}{fit_sgr model for survival growth and recruitment for the given
species.}

\item{species}{Name of the species to run simulation on. Single char.}

\item{correction}{IPM correction wanted for this species. Single char.}
}
\description{
Constructor of mutrix class
}
