% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fit_sgr.R
\name{old_fit2fit}
\alias{old_fit2fit}
\title{load old fitted models}
\usage{
old_fit2fit(species, path = here(), replicat = 42, mean = FALSE)
}
\arguments{
\item{species}{Name of the species to run simulation on. Single char.}

\item{path}{Place to save the resulting file. Single Char.}

\item{replicat}{Numeric for the simulation to select. By default, the 42th.}

\item{mean}{Should the return element be a mean of all models or a single value.
FALSE by default, TRUE will ignore replicat argument}
}
\description{
load old fitted models
}
