% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_forest.R
\name{old_ipm2forest}
\alias{old_ipm2forest}
\title{Dev function to read old ipm}
\usage{
old_ipm2forest(sp_name, climatic = 1, path = here(), replicat = 42)
}
\arguments{
\item{sp_name}{Name of the species to add in a single forest object. char.}

\item{climatic}{Vector of climatic situations to run on. IPM must exist for
each one or else this climatic value will be skipped. int.}

\item{path}{Place to save the resulting file. Single Char.}

\item{replicat}{Numeric for the simulation to select. By default, the 42th.}
}
\description{
Read old ipm and put them in the new class format.
}
