% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{old_ipm2species}
\alias{old_ipm2species}
\title{Dev function to read old ipm}
\usage{
old_ipm2species(
  species,
  climatic = 1,
  path = here(),
  replicat = 42,
  harvest = def_harv,
  init_pop = def_init
)
}
\arguments{
\item{species}{Name of the species to run simulation on. Single char.}

\item{climatic}{Vector of climatic situations to run on. IPM must exist for
each one or else this climatic value will be skipped. int.}

\item{path}{Place to save the resulting file. Single Char.}

\item{replicat}{Numeric for the simulation to select. By default, the 42th.}

\item{harvest}{Function to impact the population with harvest rule.
Argument must be \code{pop}.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most usefull when creating random
population.}
}
\description{
Read old ipm and put them in the new class format.
}
