% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_X.R
\name{print.pop_state}
\alias{print.pop_state}
\title{pop_state print method}
\usage{
\method{print}{pop_state}(x, ...)
}
\arguments{
\item{x}{pop_state class (it is a numeric vector with attributes.)}

\item{...}{Ignored}
}
\description{
pop_state print method
}
\examples{
mesh <- seq(1, 10, by = 1)
print(state_init(mesh))

}
\keyword{internal}
