% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{print.deter_sim}
\alias{print.deter_sim}
\alias{print.summary_sim}
\title{Generics for treeforce classes.}
\usage{
\method{print}{deter_sim}(x, ...)

\method{print}{summary_sim}(x, ...)
}
\arguments{
\item{x}{summary_sim class}

\item{...}{Ignored}
}
\description{
Generics for treeforce classes.
}
\keyword{internal}
