% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{run_sim_deter}
\alias{run_sim_deter}
\title{Simulate multiple scenarii for a species}
\usage{
run_sim_deter(
  species,
  init_pop,
  climatic = 1,
  tlim = 3000,
  Harv = 0.006,
  BAsup = 200,
  correction = "none",
  SurfEch = 0.03,
  delay = 0,
  path = here(),
  parallel = FALSE,
  save = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{species}{Name of the species to run simulation on. Single char.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most usefull when creating random
population.}

\item{climatic}{Vector of climatic situations to run on. IPM must exist for
each one or else this climatic value will be skipped. int.}

\item{tlim}{Number of simulation iterations (years) in the future. single int.}

\item{Harv}{Percentage of harvest in the population per year.}

\item{BAsup}{Maximum basal area per hectare. This should be inferior or equal
to the BAsup of the Model. single int.}

\item{correction}{Choice of correction of the IPM between \code{"none"}
(default) and \code{"cut"}. The second option set the last column to 0 in the
IPM so that no individual can grow outside of the defines classes.}

\item{SurfEch}{Value of plot size surface in \eqn{m^2}}

\item{delay}{Number of year delay between the recruitment of an individual
and it's inclusion in the IPM. This will enlarge the IPM and add sub diagonal
values of 1. # TODO see code{link{treeforce}{delay.ipm}}.}

\item{path}{Place to save the resulting file. Single Char.}

\item{parallel}{When multiple climatic values are given, it's possible to
run computation on multiple cores. FALSE by default.}

\item{save}{Save the simulation in an \code{.Rds} file located at}

\item{verbose}{Print message, used for debugs only. FALSE by default}
}
\value{
Dataframe (tibble class) formated in long format. Column follow this order :
\itemize{
\item{name - name of the variable. Contains state_eq, BA, N, BAeq and Neq}
\item{t - time for the specified value. BA and N follow t from 1 to tlim,
whereas equilibrium value are all a the time of equilibrium}
\item{m - mesh element for the state_eq variable.}
\item{delay - Delay setup used in the simulation}
\item{clim - Climatic condition for IPM selection.}
\item{harv - Harvest setup used in the simulation}
\item{corr - Correction setup used in the simulation.}
\item{n - Number of the simulation, if multiple IPM are provided per
climate}
\item{value - final value for the variable}
}
}
\description{
Wrapper function of the \code{sim_deter_forest} that loop on different
climatic values by loading the corresponding IPM and running the replicated
simulations.
}
\details{
When using parallel computations, the number of cores used will be the
minimum between the number of climate and the number of core of the computer
\itemize{
\item 
\enumerate{
\item 
}
}
}
