% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{sim_rdikg}
\alias{sim_rdikg}
\title{Compute RDI and Kg from simulations output}
\usage{
sim_rdikg(sim, rdi_c = NULL)
}
\arguments{
\item{sim}{simulation output table.}

\item{rdi_c}{rdic coefficients for all species. Named vector. If NULL
(default, the values are taken from matreex::rdi_coef)}
}
\description{
Compute RDI and Kg from simulations output
}
