% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{sp_rec}
\alias{sp_rec}
\title{sp recruit}
\usage{
sp_rec(x, climatic)
}
\arguments{
\item{x}{Species to get the recruitment function from}

\item{climatic}{Climate vector is needed for mu_gr object to build the
corresponding recruitment function.}
}
\description{
Get species recruitment function
}
