% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_species_maturity.R
\docType{data}
\name{species_maturity}
\alias{species_maturity}
\title{Species maturity size.}
\format{
An object of class \code{data.frame} with 27 rows and 2 columns.
}
\source{
Julien Barrere
}
\usage{
species_maturity
}
\description{
Species maturity size.
}
\keyword{datasets}
