% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_X.R
\name{state_init}
\alias{state_init}
\title{pop_state initiation}
\usage{
state_init(mesh)
}
\arguments{
\item{mesh}{all possible states of a population, based on an IPM.
Minimal and maximal values are respectively U and L, for a total number of
m states.}
}
\description{
Initiate a population state by drawing from the mesh of the related IPM
}
\examples{
mesh <- seq(1, 10, by = 1)
state_init(mesh)

}
