% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub_diag}
\alias{sub_diag}
\title{Add bandeau matrix as sub diagonal of a matrix}
\usage{
sub_diag(matrix, fill, dist = 0)
}
\arguments{
\item{matrix}{Matrix to fill in diagonal}

\item{fill}{Matrix to modify and paste inside matrix}

\item{dist}{Distance to the diagonal of the matrix. Default is 0 and}
}
\value{
Modifief matrix
}
\description{
Add bandeau matrix as sub diagonal of a matrix
}
\examples{
matrix <- matrix(0, ncol = 6, nrow = 6)
fill <- matrix(1, ncol = 6, nrow = 2)
fill[,2:4 ] <- 1:6
sub_diag(matrix, fill, dist = 0)
sub_diag(matrix, fill, dist = 2)
sub_diag(matrix, fill, dist = 6)

}
\keyword{Internal}
