% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{summary.ipm}
\alias{summary.ipm}
\title{IPM summary}
\usage{
\method{summary}{ipm}(object, ...)
}
\arguments{
\item{object}{Ipm object.}

\item{...}{Dots are dots.}
}
\description{
Additionnal argument :
\describe{
\item{fline}{skip the first line. FALSE by default.}
}
}
