% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{summary.mu_gr}
\alias{summary.mu_gr}
\title{mu_gr summary}
\usage{
\method{summary}{mu_gr}(object, ...)
}
\arguments{
\item{object}{Ipm object.}

\item{...}{Dots are dots.}
}
\description{
This function is not written yet.
}
