% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{summary.species}
\alias{summary.species}
\title{species summary}
\usage{
\method{summary}{species}(object, ...)
}
\arguments{
\item{object}{Ipm object.}

\item{...}{Dots are dots.}
}
\description{
Additionnal argument :
\describe{
\item{sum_ipm}{Print IPM summary. TRUE by default.}
}
}
