% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{summary.summary_sim}
\alias{summary.summary_sim}
\title{Summary for treeforce package objects.}
\usage{
\method{summary}{summary_sim}(object, ...)
}
\arguments{
\item{object}{summary_sim class}

\item{...}{Ignored}
}
\description{
Summary for treeforce package objects.
}
\keyword{internal}
