% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{tree_format}
\alias{tree_format}
\alias{tree_format.default}
\alias{tree_format.deter_sim}
\title{tree_format generic}
\usage{
tree_format(x)

\method{tree_format}{default}(x)

\method{tree_format}{deter_sim}(x)
}
\arguments{
\item{x}{Simulations created with sim_deter_forest}
}
\description{
Format simulation output from sim_deter_forest function to a more tidyverse
format (long format) for ggplot2 and filtering.
}
\keyword{internal}
