% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree_type.R
\docType{data}
\name{tree_type}
\alias{tree_type}
\title{Species type.}
\format{
An object of class \code{data.frame} with 28 rows and 2 columns.
}
\source{
Maxime Jaunatre
}
\usage{
tree_type
}
\description{
Species type.
}
\keyword{datasets}
