% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{Uneven_harv}
\alias{Uneven_harv}
\title{Perform the harvest Uneven}
\usage{
Uneven_harv(x, species, ...)
}
\arguments{
\item{x}{population state distribution at time t}

\item{species}{The species class object of interest to get mesh and harv_lim
values from.}

\item{...}{Variables used in this case of Uneven harvest
\describe{
\item{targetBAcut}{Basal Area to cut.}
\item{ct}{is the vector to compute BA with x (ct = Buildct(mesh, SurfEch))}
}}
}
\description{
Perform the harvest Uneven
}
