% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Yggdrasil.R
\docType{data}
\name{Yggdrasil}
\alias{Yggdrasil}
\alias{plots_pred_Yggdrasil}
\alias{fit_sgr_all_Yggdrasil}
\alias{IPM_Yggdrasil_Clim_1}
\alias{IPM_Yggdrasil_Clim_2}
\alias{IPM_Yggdrasil_Clim_3}
\title{Yggdrasil species dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 4 columns.

An object of class \code{list} of length 6.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.
}
\usage{
plots_pred_Yggdrasil

fit_sgr_all_Yggdrasil

IPM_Yggdrasil_Clim_1

IPM_Yggdrasil_Clim_2

IPM_Yggdrasil_Clim_3
}
\description{
A dataset with all neccessary elements for an old way to simulate
a species IPM along time. This will be modified as I set new classes !
}
\examples{
data(plots_pred_Yggdrasil)
data(fit_sgr_all_Yggdrasil)
data(IPM_Yggdrasil_Clim_1)
data(IPM_Yggdrasil_Clim_2)
data(IPM_Yggdrasil_Clim_3)

}
\author{
Jaunatre Maxime \href{mailto:maxime.jaunatre@yahoo.fr}{maxime.jaunatre@yahoo.fr}
}
\keyword{datasets}
