% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plod_pred.R
\docType{data}
\name{climate_species}
\alias{climate_species}
\title{Species climate optimum dataset.}
\format{
An object of class \code{data.frame} with 81 rows and 11 columns.
}
\source{
https://doi.org/10.1111/1365-2745.13533
}
\usage{
climate_species
}
\description{
Climate defined per species from Kunstler et al. 2021. Each species
had its IPM integrated for 3 climates : optimum on climate 2,
hot edge at climate 1 and cold edge for climate 3. Climate was defined
from sgdd and wai variables.
}
\keyword{datasets}
