% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{climatic}
\alias{climatic}
\alias{climatic.ipm}
\title{Get climatic values in treeforce class object}
\usage{
climatic(x)

\method{climatic}{ipm}(x)
}
\description{
Get climatic values in treeforce class object
}
