% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{climatic}
\alias{climatic}
\title{Get climatic values in treeforce class object}
\usage{
climatic(x)
}
\arguments{
\item{x}{treeforce class object. Used on ipm and species at this time.}
}
\description{
Get climatic values in treeforce class object
}
