% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_harv}
\alias{def_harv}
\title{Default population harvest}
\usage{
def_harv(x)
}
\arguments{
\item{x}{population state at time t}
}
\value{
Distribution of harvest achieved on the population.
Values are between 0 (null harvest) and Xi.
}
\description{
Default population harvest
}
