% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_harv}
\alias{def_harv}
\title{Default population harvest}
\usage{
def_harv(x, species, harv_rule, targetBA, ct, t)
}
\arguments{
\item{x}{population state at time t}

\item{species}{ignored}

\item{harv_rule}{ignored}

\item{targetBA}{ignored}

\item{ct}{ignored}

\item{t}{ignored}
}
\value{
Distribution of population to harvest.
Values are between 0 (null harvest) and Xi.
}
\description{
Default population harvest
}
