% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_harv}
\alias{def_harv}
\title{Default population harvest}
\usage{
def_harv(x, species, targetBAcut, ct)
}
\arguments{
\item{x}{population state at time t}

\item{species}{ignored}

\item{targetBAcut}{ignored}

\item{ct}{define this # TODO}
}
\value{
Distribution of population to harvest.
Values are between 0 (null harvest) and Xi.
}
\description{
Default population harvest
}
