% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init}
\alias{def_init}
\title{Default population initiation}
\usage{
def_init(mesh, SurfEch = 0.03)
}
\arguments{
\item{mesh}{all possible states of a population, based on an IPM.
Minimal and maximal values are respectively U and L, for a total number of
m states.}

\item{SurfEch}{Value of plot size surface in \eqn{m^2}}
}
\description{
Default population initiation
}
