% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{def_init_k}
\alias{def_init_k}
\title{Init population with precise distribution}
\usage{
def_init_k(x)
}
\arguments{
\item{x}{Distribution to draw systematically. This distribution should
be composed of values in \code{[0, Inf]} values with a sum superior to 0.
This is the distribution per hectare and not for the sampled plot.}
}
\value{
Function similar def_init but with no random effect anymore.
}
\description{
Init population with precise distribution
}
\details{
The resulting function will check if the provided vector is the same length
as mesh.
}
\seealso{
Other functions for initiating species population during simulation: 
\code{\link{def_init}()},
\code{\link{def_initBA}()},
\code{\link{def_init_even}()},
\code{\link{species}()}
}
\concept{functions for initiating species population during simulation}
