% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{delay}
\alias{delay}
\title{Add delay states to a system}
\usage{
delay(x, delay = 0)
}
\arguments{
\item{x}{an object that require a delay addition}

\item{delay}{the number of time delay to add. single positive int.}
}
\description{
Add delay states to an existing pop_state class object or an ipm class
object.
}
\details{
This function is a method that call \code{delay.ipm} internal function.
}
