% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{Even_harv}
\alias{Even_harv}
\title{Perform the harvest Even}
\usage{
Even_harv(x, species, ...)
}
\arguments{
\item{x}{population state distribution at time t}

\item{species}{The species class object of interest to get mesh and RDIcoef
values from. RDIcoef is a one line dataframe with RDI coefficient for one
species.}

\item{...}{Variables used in this case of Uneven harvest
\describe{
\item{targetKg}{is the target Kg}
\item{targetRDI}{is the target RDI}
\item{ct}{is the vector to compute BA with x (ct = Buildct(mesh, SurfEch))}
\item{SurfEch}{Value of plot size surface in ha}
}}
}
\description{
Perform the harvest Even
}
\details{
Even harvest is done at the stand level with the getPcutEven function
to acomodate with the multispecific case.
}
\seealso{
Other functions that defines harvest rules for a species.: 
\code{\link{Uneven_harv}()},
\code{\link{def_harv}()},
\code{\link{species}()}
}
\concept{functions that defines harvest rules for a species.}
