% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fit_sgr.R
\docType{data}
\name{fit_data}
\alias{fit_data}
\alias{fit_species}
\alias{fit_Abies_alba}
\alias{fit_Acer_campestre}
\alias{fit_Acer_pseudoplatanus}
\alias{fit_Alnus_glutinosa}
\alias{fit_Betula}
\alias{fit_Carpinus_betulus}
\alias{fit_Fagus_sylvatica}
\alias{fit_Fraxinus_excelsior}
\alias{fit_Juniperus_thurifera}
\alias{fit_Larix_decidua}
\alias{fit_Picea_abies}
\alias{fit_Pinus_halepensis}
\alias{fit_Pinus_nigra}
\alias{fit_Pinus_pinaster}
\alias{fit_Pinus_pinea}
\alias{fit_Pinus_sylvestris}
\alias{fit_Pinus_uncinata}
\alias{fit_Populus_tremula}
\alias{fit_Prunus_padus}
\alias{fit_Quercus_faginea}
\alias{fit_Quercus_ilex}
\alias{fit_Quercus_petraea}
\alias{fit_Quercus_pubescens}
\alias{fit_Quercus_pyrenaica}
\alias{fit_Quercus_robur}
\alias{fit_Quercus_suber}
\alias{fit_Salix_caprea}
\title{Species fited models dataset.}
\format{
An object of class \code{character} of length 27.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.

An object of class \code{fit_sgr} of length 4.
}
\source{
https://doi.org/10.1111/1365-2745.13533
}
\usage{
fit_species

fit_Abies_alba

fit_Acer_campestre

fit_Acer_pseudoplatanus

fit_Alnus_glutinosa

fit_Betula

fit_Carpinus_betulus

fit_Fagus_sylvatica

fit_Fraxinus_excelsior

fit_Juniperus_thurifera

fit_Larix_decidua

fit_Picea_abies

fit_Pinus_halepensis

fit_Pinus_nigra

fit_Pinus_pinaster

fit_Pinus_pinea

fit_Pinus_sylvestris

fit_Pinus_uncinata

fit_Populus_tremula

fit_Prunus_padus

fit_Quercus_faginea

fit_Quercus_ilex

fit_Quercus_petraea

fit_Quercus_pubescens

fit_Quercus_pyrenaica

fit_Quercus_robur

fit_Quercus_suber

fit_Salix_caprea
}
\description{
Reformated fitted models from Kunstler et al. 2021. Each species
is fitted for growth, survival and recruitment along climatic conditions
based on sgdd and wai. All objects were modified with the oldfit2fit function
from this package.
}
\details{
The fit_species dataset contains the list of species stored with the package.
}
\keyword{datasets}
