% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_forest.R
\name{forest}
\alias{forest}
\title{Create a new forest for simulation}
\usage{
forest(species = list())
}
\arguments{
\item{species}{List of species created with treeforce package.}
}
\description{
Only used in the treeforce package
}
