% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipm.R
\name{make_IPM}
\alias{make_IPM}
\title{Build IPM for a given species and climate}
\usage{
make_IPM(
  species,
  climate,
  clim_lab,
  fit,
  mesh = c(m = 700, L = 90, U = 1500),
  BA = 0:100,
  correction = c("constant", "none", "ceiling", "sizeExtremes"),
  level = 420,
  diag_tresh = 50,
  midbin_tresh = 100,
  IsSurv = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{species}{The species names to be registered in the object}

\item{climate}{Named vector of the environmental variables used in the fitted
model. Was data_plot_pred before.}

\item{clim_lab}{Label for climatic used. This values will be matched when
simulating multiple species together.}

\item{fit}{Fitted model for growth and survival of the species and climate
given. Functions will depend on size and basal area.}

\item{mesh}{vector of mesh variables. m is the number of bins, L is the
minimum size and U the maximum size. h will be defined in the function as
\eqn{h <- (U - L) / m}.}

\item{BA}{Vector of basal area to integrate on. Integrating on 0 is important
so use it. Integrating above 200 is absurd.}

\item{correction}{Correction to apply to the IPM matrix for eviction. Choices
constant (default), ceiling, sizeExtremes and none.}

\item{level}{TODO}

\item{diag_tresh}{Treshold}

\item{midbin_tresh}{Number of cells external to the GL integration to
integrate with the mid bin method.}

\item{IsSurv}{Adding survival to the IPM. Set to FALSE is usefull to test for
eviction of the model. TRUE by default.}

\item{verbose}{Print message. FALSE by default}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Integrate IPM for growth and survival function at a specific climate for a
species on a basal area variation.

The check between climate variables and fitted variable will assert if all
variables in the model are provided expect variables derived from "size"
(size, size2, logsize), "intercept" and "BATOTcomp". An error will be
triggered if the climate variable is missing.

One can desactivate each kind of integration with some treshold values.
A negative value in diag_tresh (ex: -1) will cancel the Gauss-Legendre
integration and a midbin_tresh null value (ex: 0) will cancel the midbin
integration.
}
