% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matreex.R
\docType{package}
\name{matreex}
\alias{matreex-package}
\alias{matreex}
\title{matreex}
\description{
This R package aims to simulate tree population dynamics with integral
projection models for single or multiple species, both including density
dependence. The simulations return the size distribution of each species
at each time step. These models can include different harvest and
disturbance events. The package includes a function to run the model until
equilibrium.This package is developped for LESSEM internal users only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forge.inrae.fr/lessem/rpackages/matreex}
}

}
\author{
\strong{Maintainer}: Maxime Jaunatre \email{maxime.jaunatre@inrae.fr} (\href{https://orcid.org/0009-0002-2816-1677}{ORCID}) [translator]

Authors:
\itemize{
  \item Julien Barrere \email{julien.barrere@inrae.fr} (\href{https://orcid.org/0000-0002-6686-726X}{ORCID})
  \item Thomas Cordonnier \email{thomas.cordonnier@onf.fr} (\href{https://orcid.org/0000-0003-3684-4662}{ORCID})
  \item Arnaud Guyennon (\href{https://orcid.org/0000-0003-2178-3801}{ORCID})
  \item Bjoern Reineking \email{bjoern.reineking@inrae.fr} (\href{https://orcid.org/0000-0001-5277-9181}{ORCID})
  \item Georges Kunstler \email{georges.kunstler@inrae.fr} (\href{https://orcid.org/0000-0002-2544-1940}{ORCID})
}

Other contributors:
\itemize{
  \item Anne Baranger (\href{https://orcid.org/0000-0001-6929-6362}{ORCID}) [contributor]
  \item Laura Touzot (\href{https://orcid.org/0000-0003-0445-554X}{ORCID}) [contributor]
  \item INRAE [copyright holder]
}

}
