% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matreex.R
\docType{package}
\name{matreex}
\alias{matreex}
\title{matreex}
\description{
This package is developped for LESSEM internal users only.
It allow to run simulation from IPM and NFI datasets.
}
\details{
Fill the package description !
}
