% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ipm.R
\name{new_ipm}
\alias{new_ipm}
\title{Constructor for IPM class}
\usage{
new_ipm(IPM, BA, mesh, species, climatic, compress = FALSE)
}
\arguments{
\item{IPM}{list of the IPM for different basal area for the species of
interest. List of dtCMatrix.}

\item{BA}{values of BA defined for the IPM, with the same names as IPM. num.}

\item{mesh}{Vector of all values between L and U of the IPM
(L being the smallest size and U the largest.). Length of this vector is
the number of size class in the IPM. num.}

\item{species}{Name of the species to run simulation on. Single char.}

\item{climatic}{Vector of climatic situations to run on. IPM must exist for
each one or else this climatic value will be skipped. int.}

\item{compress}{Is the IPM matrix compressed as integer (via \code{x * 1e7}).
Help to limit  size when saved on disc. FALSE by default. lgl.}
}
\description{
Constructor for IPM class
}
