% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{new_species}
\alias{new_species}
\title{Constructor for species class}
\usage{
new_species(
  IPM,
  init_pop,
  harvest_fun,
  disturb_fun,
  harv_lim = c(dth = 175, dha = 575, hmax = 1),
  rdi_coef = NULL,
  disturb_coef = NULL
)
}
\arguments{
\item{IPM}{ipm class object from the treeforce package.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most useful when creating random
population.}

\item{harvest_fun}{Function to impact the population with harvest rule.
Argument must be \code{x}, \code{species},  \code{...},}

\item{disturb_fun}{Function to impact the population with possibles
disturbances. Extra care is needed to give this function all needed parameters/
Default is \code{def_disturb}.
\code{BAtarget}, \code{ct} and \code{t}.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees. It represent the
distribution of the population to harvest}

\item{harv_lim}{Limits of harvest for a population size distribution.
\describe{
\item{dth}{minimum diameter at which we cut the size distribution}
\item{dha}{harvest diameter}
\item{hmax}{maximum harvest rate for a size class}
}}

\item{rdi_coef}{Coefficient for RDI curve used in even harvest.
The model require the intercept and slope.}
}
\description{
Only used in the treeforce package
}
\keyword{internal}
