% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{old_ipm2species}
\alias{old_ipm2species}
\title{Dev function to read old ipm}
\usage{
old_ipm2species(
  species,
  climatic = 1,
  path = here(),
  replicat = 42,
  harvest = def_harv,
  disturb = def_disturb,
  init_pop = def_init,
  delay = 0
)
}
\arguments{
\item{species}{Name of the species to load. Single char.}

\item{climatic}{Numeric that coded old climatic state. int.}

\item{path}{Place to load the previous IPM file. Single Char.}

\item{replicat}{Numeric for the model to select. By default, the 42th.}

\item{harvest}{Function to impact the population with harvest rule.
Argument must be \code{pop}.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees.}

\item{disturb}{Function to impact the population with possibles
disturbances. Extra care is needed to give this function all needed parameters/
Default is \code{def_disturb}.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most usefull when creating random
population.}

\item{delay}{Number of year delay between the recruitment of an individual
and it's inclusion in the IPM. This will enlarge the IPM and add sub diagonal
values of 1. See \code{\link[matreex]{delay}}.}
}
\description{
Read old ipm and put them in the new species class format.
}
