% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rdi_coef.R
\docType{data}
\name{rdi_coef}
\alias{rdi_coef}
\title{Species Rdi intercept and slope.}
\format{
An object of class \code{data.frame} with 28 rows and 3 columns.
}
\source{
https://10.12688/openreseurope.13671.2
}
\usage{
rdi_coef
}
\description{
Species Rdi intercept and slope.
}
\keyword{datasets}
