% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rdi_coef.R
\docType{data}
\name{rdi_coef}
\alias{rdi_coef}
\title{Species Rdi intercept and slope.}
\format{
An object of class \code{data.frame} with 28 rows and 4 columns.
}
\source{
Personnal contact with Patrick Vallet
}
\usage{
rdi_coef
}
\description{
Species Rdi intercept and slope.
}
\details{
Missing values for Quercus faginea from french NFI
}
\keyword{datasets}
