% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{sim_deter_forest}
\alias{sim_deter_forest}
\title{Deterministic Tree population simulation}
\usage{
sim_deter_forest(
  Forest,
  tlim = 3000,
  equil_dist = 250,
  equil_diff = 1,
  equil_time = 10000,
  harvest = c("default", "Uneven", "Even"),
  targetBA = 20,
  targetRDI = 0.9,
  targetKg = 0.9,
  climate = NULL,
  disturbance = NULL,
  correction = "none",
  SurfEch = 0.03,
  verbose = FALSE
)
}
\arguments{
\item{Forest}{Group of species that each contains IPM for deterministic
transition for \eqn{Z_{t}} state in a population to \eqn{Z_{t+1}} state.
A species is also defined with recruitment and harvest functions, please
see \code{\link[matreex]{species}} for more information.}

\item{tlim}{Number of simulation iterations (years) in the future. single int.}

\item{equil_dist}{Number of last n time for which the range difference
should not exceed \code{equil_diff} during the equilibrium research.
single int.}

\item{equil_diff}{Difference threshold of the basal area under which
equilibrium is assumed. single real.}

\item{equil_time}{Total maximum time simulation allowed in equilibrium
research. Must be higher or equal to tlim and equil_dist. single int.}

\item{harvest}{Choice of harvest rules between default, Uneven and Even.
This indicate what settings will be used. See Details.}

\item{targetBA}{BA value that is targetted by the Harvest module.
Single numeric in \eqn{m^2}.}

\item{climate}{Optional, climate matrix if climate variation along time is
needed. Climate variation rely on species created with mu_gr class objects.
This matrix require as many rows as time steps until equil_time.
If the climate does not variate, a single row can given and will be reused.}

\item{correction}{Choice of correction of the IPM between \code{"none"}
(default) and \code{"cut"}. The second option set the last column to 0 in the
IPM so that no individual can grow outside of the defines classes.}

\item{SurfEch}{Value of plot size surface in ha}

\item{verbose}{Print message. FALSE by default}
}
\value{
Data.frame with long tidyverse format : a row for each observation and a
column per variable. Columns are listed below, some may contains NA values,
as for example species when there is a non-specific variable (BA).
\describe{
\item{species}{ Name of the species.}
\item{var}{Variable of interest}
\item{time}{Time step of the simulation. If the equilibrium is the last
time in \code{tlim} input, this time will occur twice in the table.}
\item{mesh}{Mesh class number, from 1 to n class.}
\item{size}{Size corresponding to the mesh class.}
\item{equil}{Logical if this time step is the equilibrium or last step of
simulation}
\item{value}{Numeric values of the variables.}
}

The variables are :
\describe{
\item{m}{Distribution of density by mesh along time per ha.}
\item{N}{Sum of density per ha. (colSums for m)}
\item{BA}{Basal area of the population per ha}
\item{h}{Distribution of harvest density by mesh along time per ha.}
\item{H}{Sum of density per ha. (colSums for h)}
}
}
\description{
Simulate a population size state during \eqn{[1, t_{lim}]} times.
The state at time \eqn{t+1} is dependent on state at time \eqn{t} and
the projection matrix (IPM).
The simulation will run until \eqn{t_{lim}} and if the equilibrium is not reached,
it will continue. Only simulation in \eqn{[1, t_{lim}]}, and equilibrium
state are returned.
}
\details{
At each iteration, the basal area is evaluated to select the corresponding
IPM matrix.
}
