% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{sim_deter_forest}
\alias{sim_deter_forest}
\alias{sim_deter_forest.species}
\alias{sim_deter_forest.forest}
\title{Deterministic Tree population simulation}
\usage{
sim_deter_forest(
  Forest,
  tlim = 3000,
  equil_dist = 250,
  equil_diff = 1,
  equil_time = 10000,
  targetBA = 20,
  correction = "none",
  SurfEch = 0.03,
  delay = 0,
  verbose = FALSE
)

\method{sim_deter_forest}{species}(
  Forest,
  tlim = 3000,
  equil_dist = 250,
  equil_diff = 1,
  equil_time = 10000,
  targetBA = 20,
  correction = "none",
  SurfEch = 0.03,
  delay = 0,
  verbose = FALSE
)

\method{sim_deter_forest}{forest}(
  Forest,
  tlim = 3000,
  equil_dist = 250,
  equil_diff = 1,
  equil_time = 10000,
  targetBA = 20,
  correction = "none",
  SurfEch = 0.03,
  delay = 0,
  verbose = FALSE
)
}
\arguments{
\item{Forest}{Group of species that each contains IPM for deterministic
transition for \eqn{Z_{t}} state in a population to \eqn{Z_{t+1}} state.
A species is also defined with recruitment and harvest functions.
test # TODO link to class species}

\item{tlim}{Number of simulation iterations (years) in the future. single int.}

\item{equil_dist}{Number of last n time for which the range difference
should not exceed \code{equil_diff} during the equilibrium research.
single int.}

\item{equil_diff}{Difference threshold of the basal area under which
equilibrium is assumed. single real.}

\item{equil_time}{Total maximum time simulation allowed in equilibrium
research. Must be higher or equal to tlim and equil_dist. single int.}

\item{targetBA}{BA value that is targetted by the Harvest module.
Single numeric in \eqn{m^2}.}

\item{correction}{Choice of correction of the IPM between \code{"none"}
(default) and \code{"cut"}. The second option set the last column to 0 in the
IPM so that no individual can grow outside of the defines classes.}

\item{SurfEch}{Value of plot size surface in ha}

\item{delay}{Number of year delay between the recruitment of an individual
and it's inclusion in the IPM. This will enlarge the IPM and add sub diagonal
values of 1. # TODO see code{link{treeforce}{delay.ipm}}.}

\item{verbose}{Print message, used for debugs only. FALSE by default}
}
\value{
Matrix of the population states for time in \eqn{[1, t_{lim}]},
plus a last column that is the state at equilibrium.
In row are the states of the population, plus two last rows :
N, the total number of individual in the population at
time \eqn{t} and BA, the basal area of the population.
\code{colnames} are set as tn with n the time of simulation.
\code{rownames} are labelled mi with i the different state of the population.
}
\description{
Simulate a population size state during \eqn{[1, t_{lim}]} times.
The state at time \eqn{t+1} is dependent on state at time \eqn{t} and
the projection matrix (IPM).
The simulation will run until \eqn{t_{lim}} and if the equilibrium is not reached,
it will continue. Only simulation in \eqn{[1, t_{lim}]}, and equilibrium
state are returned.
}
\details{
At each iteration, the basal area is evaluated to select the corresponding
IPM matrix.
}
