% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_indiv.R
\name{sim_indiv_forest}
\alias{sim_indiv_forest}
\title{Deterministic Tree population simulation}
\usage{
sim_indiv_forest(
  Forest,
  tlim = 3000,
  climate = NULL,
  disturbance = NULL,
  harvest = 0.006,
  SurfEch = 0.03,
  verbose = FALSE
)
}
\arguments{
\item{Forest}{Group of species that each contains IPM for deterministic
transition for \eqn{Z_{t}} state in a population to \eqn{Z_{t+1}} state.
A species is also defined with recruitment and harvest functions, please
see \code{\link[matreex]{species}} for more information.}

\item{tlim}{Number of simulation iterations (years) in the future. single int.}

\item{climate}{Optional, climate matrix if climate variation along time is
needed. Climate variation rely on species created with mu_gr class objects.
This matrix require as many rows as time steps until equil_time.
If the climate does not variate, a single row can given and will be reused.}

\item{disturbance}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} parameter.}

\item{SurfEch}{Value of plot size surface in ha}

\item{verbose}{Print message. FALSE by default}
}
\value{
Data.frame with long tidyverse format : a row for each observation and a
column per variable. Columns are listed below, some may contains NA values,
as for example species when there is a non-specific variable (BA).
\describe{
\item{species}{ Name of the species.}
\item{var}{Variable of interest}
\item{time}{Time step of the simulation. If the equilibrium is the last
time in \code{tlim} input, this time will occur twice in the table.}
\item{mesh}{Mesh class number, from 1 to n class. NA for variable for individual}
\item{size}{Size corresponding to the mesh class. NA for variable for individual}
\item{equil}{Logical if this time step is the equilibrium or last step of
simulation}
\item{value}{Numeric values of the variables.}
}

The variables are :
\describe{
\item{N}{Sum of density for sampled area (SurfEch).}
\item{BAsp}{Basal area of the population per sampled area (SurfEch) and species}
\item{BAstand}{Basal area of the population per sampled area (SurfEch)
and species when excluding size class below dth. See Harvesting vignette.}
\item{H}{Sum of harvested density per sampled area (SurfEch).}
}
}
\description{
Simulate a population size state during \eqn{[1, t_{lim}]} times.
The state at time \eqn{t+1} is dependent on state at time \eqn{t} and
the projection matrix (IPM).
The simulation will run until \eqn{t_{lim}} and if the equilibrium is not reached,
it will continue. Only simulation in \eqn{[1, t_{lim}]}, and equilibrium
state are returned.
}
\details{
At each iteration, the basal area is evaluated to select the corresponding
IPM matrix.

Basic simulations input are illustrated in the main vignette.
}
