% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{sp_name}
\alias{sp_name}
\title{List the species in treeforce class object}
\usage{
sp_name(x)
}
\arguments{
\item{x}{treeforce class object. Used on ipm and species at this time.}
}
\description{
List the species in treeforce class object
}
