% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{species}
\alias{species}
\alias{harvest_fun}
\alias{init_pop}
\alias{recruit_fun}
\title{Create a new species for simulation}
\usage{
species(
  IPM,
  init_pop = def_init,
  harvest_fun = def_harv,
  disturb_fun = def_disturb,
  harv_lim = c(dth = 175, dha = 575, hmax = 1),
  rdi_coef = NULL,
  disturb_coef = NULL,
  mat_size = NULL,
  type = c("Undefined", "Broadleaf", "Coniferous")
)
}
\arguments{
\item{IPM}{ipm class object from the matreex package.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most useful when creating random
population.}

\item{harvest_fun}{Function to impact the population with harvest rule.
Argument must be \code{x}, \code{species},  \code{...},}

\item{disturb_fun}{Function to impact the population with possibles
disturbances. Extra care is needed to give this function all needed
parameters. Default is \code{def_disturb}.}

\item{harv_lim}{Limits of harvest for a population size distribution.
\describe{
\item{dth}{minimum diameter at which we cut the size distribution}
\item{dha}{harvest diameter}
\item{hmax}{maximum harvest rate for a size class}
}}

\item{rdi_coef}{Coefficient for RDI curve used in even harvest.
The model require the intercept and slope.
Fitted values are associated to this package, see details.}

\item{disturb_coef}{Species coefficient  for disturbance reaction. These
values and names are highly dependent on the disturbance function.
\code{BAtarget}, \code{ct} and \code{t}.
Fitted values are associated to this package, see details.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees. It represent the
distribution of the population to harvest}

\item{mat_size}{Size above which the individual of this species are mature
and contribute to fecundity.
Fitted values are associated to this package, see details.}

\item{type}{Type of the tree, choosing between "Broadleaf" and "Coniferous".
This value is only used during biotic disturbance with a specific disturb_fun.
This is experimental.}
}
\description{
Species are defined by an IPM which is a transition matrix from size between
t and t+1, recruitment and harvest functions (see Details). Each species has these items
defined for a given climate.
An additionnal vector of harvest parameters is required with minimal size to
harvest (dth), size above wich harvest is constant (dha).
}
\details{
A species is defined by an IPM that is an integrated prediction matrix
for growth and survival functions of the species. Since the species has other
functions defined, they are accessible and editable.
\describe{
\item{\code{init_pop}}{Function to initiate a new population. Default is
\code{\link[matreex]{def_init}}.
}
\item{\code{recruit_fun}}{Function that give a distribution for recruits.
The default is defined from models associated with the IPM
(\code{x$IPM$rec}) but it's possible to replace it. For example you can
nullify the recruitment to simulate extinction.
}
\item{\code{harvest_fun}}{Function that give harvest density distribution
when an harvest event occurs (this frequence is set at the forest scale.).
The default function is \code{\link[matreex]{def_harv}} with a constant
harvest rate of 0.6 percent. Other functions are
\code{\link[matreex]{Uneven_harv}} and \code{\link[matreex]{Even_harv}}.
}
}

Fitted values for rdi_coef, disturb_coef and mat_size are associated with
the package with the names \code{data("rdi_coef")} \code{data("disturb_coef")}
and \code{data("species_maturity")}. When using the \code{species()} function
providing \code{species(..., rdi_coef = "sp")} allow the function to load the
data from the package if the species is present in \code{data("fit_species")}.
}
\seealso{
Other functions for initiating species population during simulation: 
\code{\link{def_init}()},
\code{\link{def_initBA}()},
\code{\link{def_init_even}()},
\code{\link{def_init_k}()}

Other functions that defines harvest rules for a species.: 
\code{\link{Even_harv}()},
\code{\link{Uneven_harv}()},
\code{\link{def_harv}()}
}
\concept{functions for initiating species population during simulation}
\concept{functions that defines harvest rules for a species.}
