% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{species}
\alias{species}
\title{Create a new species for simulation}
\usage{
species(IPM, init_pop, harvest_fun, recruit_fun)
}
\arguments{
\item{IPM}{ipm class object from the treeforce package.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most useful when creating random
population.}

\item{harvest_fun}{Function to impact the population with harvest rule.
Argument must be \code{pop}.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees.}

\item{recruit_fun}{Function to recruit new individual into the population.
Argument must be \code{BATOTSP}. # TODO rename argument !
Should return a single numeric value.}
}
\description{
Only used in the treeforce package
}
