% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_species.R
\name{species}
\alias{species}
\title{Create a new species for simulation}
\usage{
species(
  IPM,
  init_pop,
  harvest_fun,
  harv_lim = c(dth = 175, dha = 575, hmax = 1)
)
}
\arguments{
\item{IPM}{ipm class object from the treeforce package.}

\item{init_pop}{Function to initiate the population at simulation start.
Arguments must be \code{mesh} and \code{SurfEch}.
Using the arguments is not mandatory, it's most useful when creating random
population.}

\item{harvest_fun}{Function to impact the population with harvest rule.
Argument must be \code{x}, \code{species},  \code{harv_rule},
\code{BAtarget}, \code{ct} and \code{t}.
Should return a population state as it's take it in input, with less
population than before. Unless you want zombie trees. It represent the
distribution of the population to harvest}

\item{harv_lim}{Limits of harvest for a population size distribution.
\describe{
\item{dth}{minimum diameter at which we cut the size distribution}
\item{dha}{harvest diameter}
\item{hmax}{maximum harvest rate for a size class}
}}
}
\description{
Species are defined by an IPM which is a transition matrix from size between
t and t+1, recruitment and harvest functions. Each species has these items
defined for a given climate.
An additionnal vector of harvest parameers is required with minimal size to
harvest (dth), size above wich harvest is constant (dha).
}
