% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{species}
\alias{species}
\alias{species.ipm}
\title{List the species in treeforce class object}
\usage{
species(x)

\method{species}{ipm}(x)
}
\description{
List the species in treeforce class object
}
