% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Deter.R
\name{tree_format}
\alias{tree_format}
\alias{tree_format.deter_sim}
\alias{tree_format.summary_sim}
\title{tree_format generic}
\usage{
tree_format(x)

\method{tree_format}{deter_sim}(x)

\method{tree_format}{summary_sim}(x)
}
\arguments{
\item{x}{any treeforce package object available.}
}
\description{
tree_format generic
}
