% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{Uneven_harv}
\alias{Uneven_harv}
\title{Perform the harvest Uneven}
\usage{
Uneven_harv(x, species, harv_rule, targetBA, ct, t)
}
\arguments{
\item{x}{population state distribution at time t}

\item{species}{The species class object of interest to get mesh and harv_lim
values from.}

\item{harv_rule}{Vector for harvest rules at the scale of the forest.
\describe{
\item{Pmax}{maximum proportion of BAcut / BA}
\item{dBAmin}{the minimum BA to perform cut}
\item{freq}{Frequence at which the harvest will be executed.}
}}

\item{targetBA}{the basal area to reach after maximum cut}

\item{ct}{define this # TODO}

\item{t}{time t of simulation}
}
\description{
Perform the harvest Uneven
}
